refactor(product-analytics): `lineGraphLogic.ts` to `tooltip-data.ts`
