+++ frontend/src/scenes/insights/views/LineGraph/PieChart.tsx
@@ -26,7 +26,7 @@ import {
     onChartClick,
     onChartHover,
 } from 'scenes/insights/views/LineGraph/LineGraph'
-import { lineGraphLogic } from 'scenes/insights/views/LineGraph/lineGraphLogic'
+import { createTooltipData } from 'scenes/insights/views/LineGraph/tooltip-data'
 
 import { areObjectValuesEmpty } from '~/lib/utils'
 import { groupsModel } from '~/models/groupsModel'
@@ -84,7 +84,6 @@ export function PieChart({
 
     let datasets = _datasets
 
-    const { createTooltipData } = useValues(lineGraphLogic)
     const { aggregationLabel } = useValues(groupsModel)
     const { highlightSeries } = useActions(insightLogic)
     const canvasRef = useRef<HTMLCanvasElement | null>(null)
@@ -216,6 +215,7 @@ export function PieChart({
 
                                 tooltipRoot.render(
                                     <InsightTooltip
+                                    
                                         seriesData={seriesData}
                                         hideColorCol={!!tooltipConfig?.hideColorCol}
                                         showHeader={false}
